requirements.txt
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
bleach==5.0.1
blinker==1.5
bokeh==2.4.3
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorcet==3.0.1
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
holoviews==1.15.2
hvplot==0.8.2
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.8.8
numpy==1.23.5
packaging==22.0
pandas==1.5.2
panel==0.14.1
param==1.12.2
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
plotly==5.11.0
protobuf==3.20.3
pyarrow==10.0.1
pyct==0.4.8
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyYAML==6.0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.2
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
webencodings==0.5.1
zipp==3.11.0
